wired up load avg into the api
* Added docker testing task
* Reworekd loadavg into top level openapi object

+++ internal/provider/system/linux/provider.go
@@ -36,5 +36,5 @@ type UptimeProvider interface {
 
 // LoadProvider is an interface that abstracts the getloadavg function.
 type LoadProvider interface {
-	GetLoadAverage() ([3]float64, error)
+	GetLoadAverage() ([3]float32, error)
 }
