+++ internal/provider/system/types.go
@@ -47,5 +47,5 @@ type UptimeProvider interface {
 
 // LoadProvider is an internal only interface that abstracts the getloadavg function.
 type LoadProvider interface {
-	GetLoadAverage() ([3]float64, error)
+	GetLoadAverage() ([3]float32, error)
 }
